Refactor source into triage_sandbox package; fix DNS crash on unanswered queries
Restructures flat src/helper.py + src/service.py into a proper
src/triage_sandbox/ package with split modules: service, report,
network, models, constants. Updates Dockerfile to build and install
a wheel, sets SERVICE_PATH to triage_sandbox.service.TriageSandbox.

Renames the project in pyproject.toml from assemblyline-service-triage-sandbox
to triage_sandbox and switches the build backend to hatchling.

Fixes a crash where a DNS flow with no response (resolved_ips absent from
dns_response) caused set_optional_argument to raise ValueError. The old code
passed `or []` to NetworkDNS.resolved_ips; empty list is rejected as not
"legitimate". Fixed by passing `or None` so the field is omitted entirely.
