+++ Dockerfile
@@ -1,20 +1,21 @@
 ARG branch=latest
 ARG base=cccs/assemblyline-v4-service-base
 
-# Builder: use uv to export pinned requirements from uv.lock
+# Builder: export pinned requirements and build the service wheel
 FROM python:3.12-slim AS builder
 COPY --from=ghcr.io/astral-sh/uv:0.11.19 /uv /bin/uv
 ENV UV_PYTHON_DOWNLOADS=0
 WORKDIR /app
+COPY pyproject.toml uv.lock ./
+COPY src/ ./src/
 RUN --mount=type=cache,target=/root/.cache/uv \
-    --mount=type=bind,source=uv.lock,target=uv.lock \
-    --mount=type=bind,source=pyproject.toml,target=pyproject.toml \
-    uv export --frozen --no-dev --no-emit-project -o requirements.txt
+    uv export --frozen --no-dev --no-emit-project -o requirements.txt && \
+    uv build --wheel -o dist/
 
 # AL4 service base
 FROM $base:$branch
 
-ENV SERVICE_PATH=service.TriageSandbox
+ENV SERVICE_PATH=triage_sandbox.service.TriageSandbox
 
 USER root
 
@@ -24,10 +25,11 @@ USER assemblyline
 
 WORKDIR /opt/al_service
 
-COPY --from=builder /app/requirements.txt ./
-RUN pip install --no-cache-dir --user -r requirements.txt
+COPY --chown=assemblyline:assemblyline --from=builder /app/requirements.txt ./
+COPY --chown=assemblyline:assemblyline --from=builder /app/dist/*.whl ./
+RUN pip install --no-cache-dir --user -r requirements.txt && \
+    pip install --no-cache-dir --user --no-deps *.whl
 
-COPY src/ ./
 COPY service_manifest.yml ./
 
 # Patch version in manifest
