+++ src/triage_sandbox/constants.py
@@ -1,0 +1,76 @@
+SANDBOX_NAME = "Triage Sandbox"
+SERVICE_NAME = "Triage"
+
+SCORE_MULTIPLY_FACTOR = 100
+
+DEFAULT_SIGNATURE_CLASSIFICATION = "TLP:CLEAR"
+
+EXTRA_CONFIG_FIELDS = [
+    "tags",
+    "rule",
+    "decoy",
+    "webinject",
+    "command_lines",
+    "listen_addr",
+    "listen_port",
+    "listen_for",
+    "shellcode",
+    "attr",
+]
+
+SUPPORTED_FILE_TYPES = [
+    "android/apk",
+    "android/dex",
+    "code/batch",
+    "code/hta",
+    "code/html",
+    "code/javascript",
+    "code/jscript",
+    "code/perl",
+    "code/ps1",
+    "code/python",
+    "code/shell",
+    "code/vbe",
+    "code/vbs",
+    "code/wsc",
+    "code/wsf",
+    "document/email",
+    "document/installer/windows",
+    "document/odt/chart",
+    "document/odt/formula",
+    "document/odt/graphics",
+    "document/odt/presentation",
+    "document/odt/spreadsheet",
+    "document/odt/text",
+    "document/odt/web",
+    "document/office/email",
+    "document/office/excel",
+    "document/office/hwp",
+    "document/office/mhtml",
+    "document/office/ole",
+    "document/office/onenote",
+    "document/office/passwordprotected",
+    "document/office/powerpoint",
+    "document/office/recoverystore",
+    "document/office/rtf",
+    "document/office/unknown",
+    "document/office/word",
+    "document/pdf",
+    "executable/linux/elf32",
+    "executable/linux/elf64",
+    "executable/mach-o",
+    "executable/windows/arm/dll64",
+    "executable/windows/arm/pe64",
+    "executable/windows/com",
+    "executable/windows/dll32",
+    "executable/windows/dll64",
+    "executable/windows/dos",
+    "executable/windows/ia/dll64",
+    "executable/windows/ia/pe64",
+    "executable/windows/pe",
+    "executable/windows/pe32",
+    "executable/windows/pe64",
+    "java/jar",
+    "shortcut/web",
+    "shortcut/windows",
+]
