+++ src/triage_sandbox/service.py
@@ -8,10 +8,7 @@ from assemblyline.odm.models.ontology.results import Process as ProcessModel
 from assemblyline.odm.models.ontology.results import Sandbox as SandboxModel
 from assemblyline.odm.models.ontology.results import Signature as SignatureModel
 from assemblyline.odm.models.ontology.results.malware_config import MalwareConfig
-from assemblyline_service_utilities.common.dynamic_service_helper import (
-    OntologyResults,
-    extract_iocs_from_text_blob,
-)
+from assemblyline_service_utilities.common.dynamic_service_helper import OntologyResults, extract_iocs_from_text_blob
 from assemblyline_v4_service.common.base import ServiceBase
 from assemblyline_v4_service.common.request import ServiceRequest
 from assemblyline_v4_service.common.result import Result, ResultSection, ResultTableSection
@@ -19,13 +16,19 @@ from retrying import RetryError, retry
 from triage import Client as TriageClient
 from triage.client import ServerError
 
-from helper import TriageResult
+from .constants import SUPPORTED_FILE_TYPES
+from .report import TriageResult
 
 _PROCESS_MODEL_FIELDS = frozenset(ProcessModel.fields())
 
+TRIAGE_POLL_DELAY = 10
+try:
+    MAX_ANALYSIS_TIMEOUT = int(os.environ["MAX_ANALYSIS_TIMEOUT"])
+except KeyError:
+    MAX_ANALYSIS_TIMEOUT = 600
 
-def _filter_process_prims(prims: dict) -> dict:
-    """Strip fields unknown to the ODM ProcessModel from a process primitives dict."""
+
+def _filter_process_prims(prims: dict) -> dict:  # type: ignore[type-arg]
     return {k: v for k, v in prims.items() if k in _PROCESS_MODEL_FIELDS}
 
 
@@ -43,81 +46,10 @@ def _attach_dynamic_ontology(service: ServiceBase, ontres: OntologyResults) -> N
         service.ontology.add_result_part(cast(Any, NetworkConnectionModel), nc_prims)
 
 
-# Ontology Result Constants
-SANDBOX_NAME = "Triage Sandbox"
-SERVICE_NAME = "TriageSandbox"
-
-# Derived from: https://tria.ge/docs/cloud-api/filetypes/
-# TODO: support archive files
-SUPPORTED_FILE_TYPES = [
-    "android/apk",
-    "android/dex",
-    "code/batch",
-    "code/hta",
-    "code/html",
-    "code/javascript",
-    "code/jscript",
-    "code/perl",
-    "code/ps1",
-    "code/python",
-    "code/shell",
-    "code/vbe",
-    "code/vbs",
-    "code/wsc",
-    "code/wsf",
-    "document/email",
-    "document/installer/windows",
-    "document/odt/chart",
-    "document/odt/formula",
-    "document/odt/graphics",
-    "document/odt/presentation",
-    "document/odt/spreadsheet",
-    "document/odt/text",
-    "document/odt/web",
-    "document/office/email",
-    "document/office/excel",
-    "document/office/hwp",
-    "document/office/mhtml",
-    "document/office/ole",
-    "document/office/onenote",
-    "document/office/passwordprotected",
-    "document/office/powerpoint",
-    "document/office/recoverystore",
-    "document/office/rtf",
-    "document/office/unknown",
-    "document/office/word",
-    "document/pdf",
-    "executable/linux/elf32",
-    "executable/linux/elf64",
-    "executable/mach-o",
-    "executable/windows/arm/dll64",
-    "executable/windows/arm/pe64",
-    "executable/windows/com",
-    "executable/windows/dll32",
-    "executable/windows/dll64",
-    "executable/windows/dos",
-    "executable/windows/ia/dll64",
-    "executable/windows/ia/pe64",
-    "executable/windows/pe",
-    "executable/windows/pe32",
-    "executable/windows/pe64",
-    "java/jar",
-    "shortcut/web",
-    "shortcut/windows",
-]
-
-TRIAGE_POLL_DELAY = 10
-try:
-    MAX_ANALYSIS_TIMEOUT = int(os.environ["MAX_ANALYSIS_TIMEOUT"])
-except KeyError:
-    MAX_ANALYSIS_TIMEOUT = 600  # Default to 600 seconds
-
-
-def _is_submission_not_reported(submission: dict | None) -> bool:
+def _is_submission_not_reported(submission: dict | None) -> bool:  # type: ignore[type-arg]
     if not submission:
         return True
-    else:
-        return submission.get("status", "none") != "reported"
+    return submission.get("status") != "reported"
 
 
 def _retry_on_not_found(exception: Exception) -> bool:
@@ -131,18 +63,19 @@ def _retry_on_not_found(exception: Exception) -> bool:
     retry_on_result=_is_submission_not_reported,
     retry_on_exception=_retry_on_not_found,
 )
-def wait_for_submission(service, submission_id):
+def wait_for_submission(service: Any, submission_id: str) -> dict:  # type: ignore[type-arg]
     submission = service.client.sample_by_id(submission_id)
     service.log.info(f"{submission['id']} status: {submission['status']}")
     return submission
 
 
 class TriageSandbox(ServiceBase):
-    def __init__(self, config=None):
-        super(TriageSandbox, self).__init__(config)
-        self.web_url = self.config.get("root_url").replace("api.", "").replace("/api", "").rstrip("/")
+    def __init__(self, config: dict | None = None) -> None:
+        super().__init__(config)
+        root_url: str = self.config.get("root_url") or ""
+        self.web_url = root_url.replace("api.", "").replace("/api", "").rstrip("/")
 
-    def search_triage(self, request: ServiceRequest):
+    def search_triage(self, request: ServiceRequest) -> dict | None:  # type: ignore[type-arg]
         submission = None
         try:
             if request.task.fileinfo.uri_info and request.get_param("submit_as_url"):
@@ -154,7 +87,7 @@ class TriageSandbox(ServiceBase):
             self.log.debug(f"Existing sample not found: {request.sha256}")
         return submission
 
-    def submit_triage(self, request: ServiceRequest):
+    def submit_triage(self, request: ServiceRequest) -> dict | None:  # type: ignore[type-arg]
         submission = None
         if request.task.fileinfo.uri_info and request.get_param("submit_as_url"):
             submission = self.client.submit_sample_url(url=request.task.fileinfo.uri_info.uri)
@@ -168,11 +101,7 @@ class TriageSandbox(ServiceBase):
                 )
         return submission
 
-    def start(self):
-        # ==================================================================
-        # Startup actions:
-        #   Your service might have to do some warming up on startup to make things faster
-        # ==================================================================
+    def start(self) -> None:
         self.log.info(f"start() from {self.service_attributes.name} service called")
 
     def execute(self, request: ServiceRequest) -> None:
@@ -193,9 +122,6 @@ class TriageSandbox(ServiceBase):
             if not submission:
                 self.log.info("File not found or submitted. Returning nothing.")
                 return None
-            # If you don't have a submission by now, just return nothing.
-            if not submission:
-                return None
         except ServerError as e:
             self.log.error(f"Triage Server Error: {e.status} - {e.kind} - {e.message}")
             raise
@@ -206,8 +132,8 @@ class TriageSandbox(ServiceBase):
         try:
             wait_for_submission(service=self, submission_id=submission["id"])
             triage_result = TriageResult(self.client, self.client.sample_by_id(sample_id=submission["id"]))
-            for i in triage_result.malware_config:
-                self.ontology.add_result_part(cast(Any, MalwareConfig), i.as_primitives(strip_null=True))
+            for mc in triage_result.malware_config:
+                self.ontology.add_result_part(cast(Any, MalwareConfig), mc.as_primitives(strip_null=True))
             result = Result()
             sandbox_section = ResultSection("Sandbox Information")
             sandbox_section.add_line(f"URL: {self.web_url}/{triage_result.sample.id}")
@@ -220,11 +146,11 @@ class TriageSandbox(ServiceBase):
                 process_tree = task.ontology.get_process_tree_result_section()
                 process_tree.auto_collapse = True
                 sigs_section = ResultSection(title_text="Signatures", auto_collapse=True)
-                sig_subsections = {}
+                sig_subsections: dict[str, ResultSection] = {}
                 for sig in task.ontology.get_signatures():
                     sig = cast(Any, sig)
                     name = sig.name.upper()
-                    if sig_subsections.get(name, False):
+                    if name in sig_subsections:
                         for attr in sig.attributes:
                             if attr.source.ontology_id.startswith("process_"):
                                 sig_subsections[name].add_tag(tag_type="dynamic.processtree_id", value=attr.source.tag)
@@ -233,23 +159,23 @@ class TriageSandbox(ServiceBase):
                         s.add_tag(tag_type="dynamic.signature.name", value=name)
                         for f in sig.malware_families:
                             s.add_tag(tag_type="attribution.family", value=f)
-                        if not s.heuristic:
-                            if sig.score == 1000:
-                                s.set_heuristic(5, signature=name)
-                            elif sig.score >= 800 and sig.score < 1000:
-                                s.set_heuristic(4, signature=name)
-                            elif sig.score >= 500 and sig.score < 800:
-                                s.set_heuristic(3, signature=name)
-                            elif sig.score >= 100 and sig.score < 500:
-                                s.set_heuristic(2, signature=name)
-                            else:
-                                s.set_heuristic(1, signature=name)
+                        score = sig.score
+                        if score >= 1000:
+                            s.set_heuristic(5, signature=name)
+                        elif score >= 800:
+                            s.set_heuristic(4, signature=name)
+                        elif score >= 500:
+                            s.set_heuristic(3, signature=name)
+                        elif score >= 100:
+                            s.set_heuristic(2, signature=name)
+                        else:
+                            s.set_heuristic(1, signature=name)
                         for attr in sig.attributes:
                             if attr.source.ontology_id.startswith("process_"):
                                 s.add_tag(tag_type="dynamic.processtree_id", value=attr.source.tag)
                         sig_subsections[name] = s
-                for i in reversed(sorted(sig_subsections.values(), key=lambda value: value.heuristic.heur_id)):
-                    sigs_section.add_subsection(i)
+                for sig_section in reversed(sorted(sig_subsections.values(), key=lambda s: s.heuristic.heur_id)):
+                    sigs_section.add_subsection(sig_section)
                 task_section.add_subsection(process_tree)
                 task_section.add_subsection(sigs_section)
                 ioc_section = ResultTableSection(title_text="Network IOCs", auto_collapse=True)
@@ -260,14 +186,14 @@ class TriageSandbox(ServiceBase):
                 if task.analysis.get("ttp"):
                     ttp_section = ResultSection("ATT&CK Techniques", auto_collapse=True)
                     for t in task.analysis["ttp"]:
-                        s = ResultSection(title_text=t)
-                        s.set_heuristic(10, attack_id=t)
-                        ttp_section.add_subsection(s)
+                        ttp_sub = ResultSection(title_text=t)
+                        ttp_sub.set_heuristic(10, attack_id=t)
+                        ttp_section.add_subsection(ttp_sub)
                     task_section.add_subsection(ttp_section)
                 if task.extracted:
                     malware_section = ResultSection(title_text="Malware Config", auto_collapse=True)
                     for e in task.extracted:
-                        if e.get("config", {}).get("c2", False):
+                        if e.get("config", {}).get("c2"):
                             m = ResultTableSection(title_text=f"{e['config']['family'].upper()}")
                             extract_iocs_from_text_blob(blob=json.dumps(e["config"]), result_section=m)
                             m.set_heuristic(100, signature=e["config"]["family"].upper())
@@ -281,7 +207,7 @@ class TriageSandbox(ServiceBase):
                                 )
                             )
                             malware_section.add_subsection(m)
-                    if len(malware_section.subsections) > 0:
+                    if malware_section.subsections:
                         task_section.add_subsection(malware_section)
                 sandbox_section.add_subsection(task_section)
                 if request.get_param("extract_pcap"):
@@ -290,8 +216,8 @@ class TriageSandbox(ServiceBase):
                             method="GET", path=f"/v0/samples/{triage_result.sample.id}/{task.task_id}/dump.pcapng"
                         )
                         fd, temp_path = tempfile.mkstemp(dir=self.working_directory)
-                        with os.fdopen(fd, "wb") as f:
-                            f.write(pcap)
+                        with os.fdopen(fd, "wb") as fh:
+                            fh.write(pcap)
                         request.add_extracted(
                             path=temp_path,
                             name=f"{triage_result.sample.id}-{task.task_id}-dump.pcapng",
@@ -300,37 +226,37 @@ class TriageSandbox(ServiceBase):
                     except Exception as e:
                         self.log.error(e)
                 if task.dumped and (request.get_param("extract_memdump") or request.get_param("extract_dropped_files")):
-                    for i in task.dumped:
-                        if i["kind"] == "region" and request.get_param("extract_memdump"):
-                            self.log.debug(f"Downloading{i['name']}")
+                    for dump in task.dumped:
+                        if dump["kind"] == "region" and request.get_param("extract_memdump"):
+                            self.log.debug(f"Downloading {dump['name']}")
                             try:
-                                file = self.client._req_file(
+                                file_data = self.client._req_file(
                                     method="GET",
-                                    path=f"/v0/samples/{triage_result.sample.id}/{task.task_id}/{i['name']}",
+                                    path=f"/v0/samples/{triage_result.sample.id}/{task.task_id}/{dump['name']}",
                                 )
                                 fd, temp_path = tempfile.mkstemp(dir=self.working_directory)
-                                with os.fdopen(fd, "wb") as f:
-                                    f.write(file)
+                                with os.fdopen(fd, "wb") as fh:
+                                    fh.write(file_data)
                                 request.add_extracted(
                                     path=temp_path,
-                                    name=i["name"],
+                                    name=dump["name"],
                                     description=f"Memdump file from task {triage_result.sample.id}-{task.task_id}",
                                 )
                             except Exception as e:
                                 self.log.error(e)
-                        if i["kind"] == "martian" and request.get_param("extract_dropped_files"):
-                            self.log.debug(f"Downloading{i['name']}")
+                        if dump["kind"] == "martian" and request.get_param("extract_dropped_files"):
+                            self.log.debug(f"Downloading {dump['name']}")
                             try:
-                                file = self.client._req_file(
+                                file_data = self.client._req_file(
                                     method="GET",
-                                    path=f"/v0/samples/{triage_result.sample.id}/{task.task_id}/{i['name']}",
+                                    path=f"/v0/samples/{triage_result.sample.id}/{task.task_id}/{dump['name']}",
                                 )
                                 fd, temp_path = tempfile.mkstemp(dir=self.working_directory)
-                                with os.fdopen(fd, "wb") as f:
-                                    f.write(file)
+                                with os.fdopen(fd, "wb") as fh:
+                                    fh.write(file_data)
                                 request.add_extracted(
                                     path=temp_path,
-                                    name=i["name"],
+                                    name=dump["name"],
                                     description=f"Dropped file from task {triage_result.sample.id}-{task.task_id}",
                                 )
                             except Exception as e:
