+++ tests/test_malware_config.py
@@ -1,8 +1,8 @@
 """
-Unit tests for helper.Config, helper.Ransom, and helper.Credentials -> MalwareConfig.
+Unit tests for triage_sandbox.models: Config, Ransom, and Credentials -> MalwareConfig.
 """
 
-from helper import Config, Credentials, Ransom
+from triage_sandbox.models import Config, Credentials, Ransom
 
 # ---------------------------------------------------------------------------
 # Config tests
